Shrdlite.AStar

Dataypes

data AStarState

The current state of the algoritm. Closed nodes are visited, open is a queue of nodes to visit, pathCost keeps track of the cost of getting to each node and parent is a map from each node to its parent or previous node.

Functions

aStar

Arguments

:: (WorldHolding -> Set WorldHolding)
Graph to search through

-> (WorldHolding -> Int)
Heuristic distance to the goal

-> (WorldHolding -> Bool)
Goal check function

-> WorldHolding
Starting point

-> Maybe Plan
Resulting path if it exists

Find the best path from the initial node to (one of) the goal state(s). Gives this in the form of a Plan.

aStar'

Arguments

:: AStarState
Initial state

-> (WorldHolding -> Set WorldHolding)
Graph to search through

-> (WorldHolding -> Int)
Heuristic distance to the goal

-> (WorldHolding -> Bool)
Goal check function

-> Maybe Plan
Resulting path if it exists

Much like aStar, but also takes an initial state

resultList :: WorldHolding -> Map WorldHolding WorldHolding -> Plan

Given an final node and a list Map from each node to its parent, make a Plan

stateTrans :: World -> World -> Int -> Plan

Given two World states and a counter (preferrably 0) make a Plan

updateState

Arguments

:: (WorldHolding -> Int)
Heuristic function

-> WorldHolding
Previous node

-> WorldHolding
Current node

-> AStarState
State to update

-> AStarState
New state

Update the state with a new node, unless it is already in open with a cheaper cost