Shrdlite.AmbiguityResolver

Documentation

type Ambiguity = ReaderT State (ErrorT ShrdliteError Identity)

unAmbiguity :: State -> Ambiguity a -> Either ShrdliteError a

resolveAmbiguity :: State -> [[Goal]] -> Either ShrdliteError Goal

resolveAmbiguity' :: State -> [Goal] -> Either ShrdliteError Goal

computation :: [Goal] -> Ambiguity Goal

pickOne :: [Goal] -> Ambiguity Goal

Picks one goal if possible. TODO: questions

sameGoalType :: Goal -> Goal -> Ambiguity Goal

Throws an OtherError if the goals do not have the same type

addObjects :: Goal -> (Set Id, Set Id) -> (Set Id, Set Id)

Adds the object Ids from the Goal to the Sets

findThe :: Goal -> (Bool, Bool) -> (Bool, Bool)

Check if there is a The quantifier in either object in the Goal or if the Bool tuple has True in the corresponding field

findAmbiguity :: Bool -> Set Id -> [Goal] -> Ambiguity ()

If the first argument is False, there is no ambiguity. Otherwise, there is an amibiguity if there are serveral Ids in the set, and then an ShrdliteError is thrown

showObjects :: Objects -> Set Id -> String

Finds the objects in the set and shows them in an appropriate manner

showObjectHelper :: Bool -> Bool -> [Object] -> String

getObjectAttribs :: [Object] -> [(Size, Color)]

allEqual :: Eq a => [a] -> Bool

validGoals :: [Goal] -> Ambiguity [Goal]

Gets all the possible goals

possibleGoal :: Goal -> Ambiguity Bool

Checks if the provided goal is possible.