Shrdlite.Common

Types

type Utterance = [String]

type Id = String

type Objects = Map Id Object

type Plan = [String]

type World = [[Id]]

type WorldChange = (Int, [Id])

type WorldHolding = (World, Maybe Id)

Datatypes

data Goal

Constructors

TakeGoal GoalObject

MoveGoal Relation GoalObject GoalObject

data GoalObject

Constructors

Flr

Obj Quantifier Id

data ValidatedGoal

Constructors

SingleGoal Goal

OrGoal [Goal]

AndGoal [Goal]

data State

Constructors

State

Fields

world :: World

holding :: Maybe Id

objects :: Objects

data ShrdliteError

Data structure for different errors that can show up during a Shrdlite run

Constructors

EntityError String

AmbiguityError [Goal] String

OtherError String

Functions

shrdliteErrorMsg :: ShrdliteError -> String

Gets the message from en error

isAmbiguityError :: ShrdliteError -> Bool

Checks wether the provided error is an AmbiguityError.

findObjPos :: Id -> World -> Maybe (Int, Int)

Finds the column and height of the object with the provided id

maybeOk :: Result a -> Maybe a

Converts a Result into a Maybe Returns Nothing in case of Error otherwise it returns Just value

ok :: Result a -> a

Unwrapper for the result type

parseObjects :: JSObject JSValue -> Objects

Extracts an Object from a JSObject

validate :: Object -> Object -> Bool

Tests whether the first object can be placed immediately above the second object without violating the given constraints.

validateAllLaws :: Relation -> Object -> Object -> Bool

Validates all physical laws. This does not look if the objects exists in the world, and it does not care about their locations.

getSize :: Object -> Size

getColor :: Object -> Color

getForm :: Object -> Form