Shrdlite.Interpreter

Documentation

type Interpretation = ReaderT State (ErrorT ShrdliteError Identity)

The Interpretation monad. Errors in the interpretation are propagated.

unInterpret :: State -> Interpretation a -> Either ShrdliteError a

Computes the interpretation with the given world state.

isEntityError :: ShrdliteError -> Bool

Checks wether the provided error is an EntityError.

interpretAll :: State -> [Command] -> Either ShrdliteError [[Goal]]

Interprets all the provided commands in the given state.

interpret :: State -> Command -> Either ShrdliteError [Goal]

Converts a parse tree into a PDDL representation of the final goal of the command

interpret' :: Command -> Interpretation [Goal]

Takes a command and tries to interpret it into a PDDL representation.

takeEntity :: Entity -> Interpretation [Goal]

Generates a list of TakeGoals with the provided Entity.

dropAtLocation :: Location -> Interpretation [Goal]

Generates a list of MoveGoals with the provided Location, using the currently held item as the entity to move.

moveEntity :: Entity -> Maybe Location -> Interpretation [Goal]

Generates a list of MoveGoals with the provided Entity and Location.

searchObjects :: Object -> Quantifier -> Maybe Location -> Interpretation [(Id, Object)]

Searches the world state after objects matching the provided quantifier and location.

locationHolds :: (Id, Object) -> Location -> Interpretation Bool

Checks if a location holds for an object in the world.

findObjColumn :: Id -> Interpretation Int

Finds the column index of the object whose Id matches the provided one.

findObjHeight :: Id -> Interpretation Int

Finds the height index of the object whose Id matches the provided one.

sameColumn :: Id -> Id -> Interpretation Bool

Checks if two objects are in the same column.

inside :: Id -> Id -> Interpretation Bool

Checks if the first object is inside the second.

ontop :: Id -> Id -> Interpretation Bool

Checks if the first object is on top of the second.

above :: Id -> Id -> Interpretation Bool

Checks if the first object is above the second.

under :: Id -> Id -> Interpretation Bool

Checks if the first object is under the second.

leftof :: Id -> Id -> Interpretation Bool

Checks if the first object is to the left of the second.

rightof :: Id -> Id -> Interpretation Bool

Checks if the first object is to the right of the second.

findSingleEntity :: Entity -> Interpretation Id

Finds a single entity

findEntity :: Entity -> Interpretation [Id]

Finds all object ids matching the entity in the provided state

column :: State -> Id -> Maybe [Id]

Returns the column index of the object id, if any.

getQuantifier :: Entity -> Quantifier

Finds the quantifier from an entity