Shrdlite.Planner

Documentation

solve :: World -> Maybe Id -> Objects -> [Goal] -> Maybe Plan

Creates Just a list of moves which together creates a Plan. The plan can consist of messages to the user and commands in the form of "pick FLOOR_SPACE" and "drop FLOOR_SPACE". If no plan could be found, returns Nothing

simulatePlan :: World -> Maybe Id -> Plan -> (World, Maybe Id)

worldGraph :: Objects -> WorldHolding -> Set WorldHolding

Given a state, produces a set of all possible neighbours

placeObject :: Objects -> Id -> ([[Id]], [[Id]]) -> Set WorldHolding -> Set WorldHolding

Puts the World parts together, with the Id added to the beginning of the second list

takeObject :: Objects -> ([[Id]], [[Id]]) -> Set WorldHolding -> Set WorldHolding

Takes the top object from the first list of the last World part

takeHighest :: ([[Id]], [[Id]]) -> Maybe Id

Takes the highest object in one of the lists

maybeLast :: [a] -> Maybe a

The last element of the list, if it is not empty

joinModified :: Objects -> ([[Id]], [[Id]]) -> ([Id] -> [Id]) -> Maybe [[Id]]

Joins two parts of the world together, modifying the first element in the second list. The function is what modifies the element (which is itself a list) and is normally either placing or taking an object from the top of the list. Returns Just a new world

val :: Objects -> [Id] -> Maybe [Id]

Takes a column and checks if the the two topmost objects are in the correct order. If there are less than two object we assume there can be no conflicts.

heuristics :: Goal -> WorldHolding -> Int

Heuristic defining how good a state is

cheapestCost :: Id -> Id -> World -> Int

Returns the cheapest heuristics to move all objects above one of the two given objects.

getFloorSpace :: World -> Bool

Checks whether there exits an empty floor space.

makeFloorSpace :: World -> [Int]

Returns the number of elements in every column.

clmn :: Int -> Id -> World -> Maybe Int

Returns the column a given object exists in.

idHeight :: Id -> World -> Maybe Int

Gives an int for how far down the object is

idHeight' :: Id -> [Id] -> Maybe Int

A helper function to idHeight which returns the number of objects above a given object if it exits.

objPos :: Id -> World -> Maybe Int

Returns an objects height in a column where 0 is the bottom.

isOver

Arguments

:: Id
Object above

-> Id
Object below

-> Bool
Should the object be just above?

-> World
World to look in

-> Bool

Checks if the first object is over the second.

check :: Goal -> WorldHolding -> Bool

Checks if the goal is fulfilled in the provided state

isBeside :: Id -> Id -> World -> Bool

Checks if the second object is in the column strictly beside the first objects column.